change the ai generative image prompt and modify the openai helper
