+++ src/components/coauthor/DraftComp.tsx
@@ -6,8 +6,7 @@ import TextareaAutosize from "@mui/material/TextareaAutosize";
 import React, { useEffect, useState } from "react";
 
 import { DESIGN_SYSTEM_COLORS } from "@/lib/theme/colors";
-
-import { sendMessageToChatGPT } from "../../services/openai";
+import { sendMessageToChatGPT } from "@/pages/api/openAI/helpers";
 
 interface Props {
   theme: any;
